refactor: 유효성 검증 구조 개선
