--- src/types/validationError.ts
@@ -1,4 +1,0 @@
-export interface ValidationError {
-  field: string;
-  message: string;
-}
